Move the unequal-depth padding policy into padding.rs
Batching fracadd trees of unequal depths needs three things: a table of
prefix equality weights over the padding coordinates, one padded layer
prover per tree, and the inverse map that turns a padded leaf claim back
into a claim on the tree's own witness. All three lived in mod.rs beside
the layer driver, which is a different concern.

They move to fracaddcheck/padding.rs: PaddedLayerProver, layer_provers,
unpad_leaf_claim, and the prefix table with its batch inversion.

The moved code is verbatim. The only signature change is visibility:
PaddedLayerProver and layer_provers become pub(super), and
unpad_leaf_claim stays pub, re-exported from mod.rs so
fracaddcheck::unpad_leaf_claim still resolves. logup_star/prove.rs
imports it by that path and compiles untouched.

The new module doc states the padding identity once, with the N' and D'
formulas and why the prover never materializes a padded witness. Two
places in mod.rs that restated it now point at the module instead.

reduce_layer stays where it is. Splitting that is a separate concern.

A struct over the pad-length bookkeeping was considered and rejected. The
property that matters is that pad_lens, provers and claims stay index
aligned, and a newtype over Vec<usize> does not enforce that; the prefix
table and its inversion are function-local temporaries, not shared state.
It would buy naming, not an invariant.
